essai comme av

+++ woody_woodpacker/includes/woody.h
@@ -6,7 +6,7 @@
 /*   By: tktorza <tktorza@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/14 16:37:57 by tktorza           #+#    #+#             */
-/*   Updated: 2017/11/24 11:47:41 by tktorza          ###   ########.fr       */
+/*   Updated: 2017/11/24 13:01:45 by tktorza          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ void    *open_decrypt(struct stat *buf, int *fd/*, int *gap*/);
 void    woody_start(void *ptr, unsigned int size, int fd);
 Elf64_Phdr *elf_find_gap(void *ptr, int size, int *p, int *len);
 Elf64_Shdr *elf_find_section(void *ptr, char *name);
-int		elf_mem_subst(void *m, int len, long pat, long val);
+int		elf_mem_subst(void *m, int len, long pat, unsigned long long val);
 void	open_woody(void *ptr, unsigned int size, int fd1, int fd2);
